added replace confirm dialog when saving

+++ api/upload.php
@@ -1,8 +1,10 @@
 <?php
 
-echo "----" . PHP_EOL;
-print_r($_FILES);
-echo "----" . PHP_EOL . PHP_EOL;
+// TODO rewrite to json response with error codes
+
+// echo "----" . PHP_EOL;
+// print_r($_FILES);
+// echo "----" . PHP_EOL . PHP_EOL;
 
 switch ($_POST['type']) {
     case "sound":
@@ -16,14 +18,20 @@ switch ($_POST['type']) {
         break;
 }
 
+$overwrite = array_key_exists("overwrite", $_POST) && ($_POST["overwrite"] === "true");
 
 mkdir($target_folder, 0777, true);
 
 foreach ($_FILES['file']['error'] as $index => $err) {
     $new_name = $target_folder . $_FILES['file']['name'][$index];
     $tmp_name = $_FILES['file']['tmp_name'][$index];
-    echo $tmp_name . '-> ' . $new_name . PHP_EOL;
+    // echo $tmp_name . '-> ' . $new_name . PHP_EOL;
     if ($err == UPLOAD_ERR_OK) {
+
+        if (file_exists($new_name) && !$overwrite) {
+            echo 'file exists' . PHP_EOL;
+            continue;
+        }
         move_uploaded_file($tmp_name, $new_name);
         echo 'ok' . PHP_EOL;
     } else {
